+++ extra/static/owncloud.js
@@ -11,13 +11,36 @@
 // This file is loaded in WebRTC context
 
 "use strict";
-define([
+
+// TODO(leon): :(
+(function() {
+var modules = [
 	'angular',
 	'moment',
 	'./PostMessageAPI.js',
-	'./config/OwnCloudConfig.js',
-], function(angular, moment, PostMessageAPI, OwnCloudConfig) {
+];
+// TODO(leon): Create helper script with this function, as we also need it in webrtc.js
+var getQueryParam = function(param) {
+	var query = window.location.search.substring(1);
+	var vars = query.split("&");
+	for (var i = 0; i < vars.length; i++) {
+		var pair = vars[i].split("=");
+		if (pair[0] === param) {
+			return window.decodeURIComponent(pair[1]);
+		}
+	}
+	return false;
+};
+if (getQueryParam('load_config_js') !== false) {
+	modules.push('./config/OwnCloudConfig.js');
+}
+// Make sure OwnCloudConfig is always the last argument
+define(modules, function(angular, moment, PostMessageAPI, OwnCloudConfig) {
 	'use strict';
+
+	if (typeof OwnCloudConfig === 'undefined') {
+		OwnCloudConfig = {OWNCLOUD_ORIGIN: '',};
+	}
 
 	var HAS_PARENT = window !== parent;
 	// TODO(leon): Create helper script with this function, as we also need it in webrtc.js
@@ -135,6 +158,7 @@ define([
 					}
 				})();
 
+				// TODO(leon): Remove is* values
 				var online = ownCloud.deferreds.online;
 				var isOnline = false;
 				var admin = ownCloud.deferreds.admin;
@@ -394,6 +418,13 @@ define([
 
 			}]);
 
+			// See User::getSignedCombo()
+			app.filter("displayUserid", [function() {
+				return function(id) {
+					return id.replace(/\|/g, ":");
+				};
+			}]);
+
 			app.directive("settingsAccount", [function() {
 				return {
 					scope: false,
@@ -407,7 +438,7 @@ define([
 				};
 			}]);
 
-			app.directive("roomBar", ["$window", "$timeout", "ownCloud", "alertify", function($window, $timeout, ownCloud, alertify) {
+			app.directive("roomBar", ["$window", "$q", "$timeout", "ownCloud", "alertify", function($window, $q, $timeout, ownCloud, alertify) {
 				var open = function() {
 					/*var popup = $window.open(
 						ownCloud.getConfig().baseURL + "/admin/tp",
@@ -464,8 +495,13 @@ define([
 					link: function(scope, element) {
 						// Hide roombar
 						//element.hide();
-						ownCloud.deferreds.admin.promise.then(function() {
-							addGenerateTemporaryPasswordButton(element);
+						$q.all({
+							"isAdmin": ownCloud.deferreds.admin.promise,
+							"isTemporaryPasswordFeatureEnabled": ownCloud.deferreds.features.temporaryPassword.promise,
+						}).then(function(args) {
+							if (args.isAdmin && args.isTemporaryPasswordFeatureEnabled) {
+								addGenerateTemporaryPasswordButton(element);
+							}
 						});
 					}
 				};
@@ -844,3 +880,4 @@ define([
 	}
 
 });
+})();