+++ helper/helper.php
@@ -11,7 +11,7 @@
 
 namespace OCA\SpreedME\Helper;
 
-use OCA\SpreedME\Config\Config;
+use OCA\SpreedME\Errors\ErrorCodes;
 use OCA\SpreedME\Settings\Settings;
 
 class Helper {
@@ -49,6 +49,53 @@ class Helper {
 		return $protocol . '://' . $hostname;
 	}
 
+	private static function getFileConfigValue($key) {
+		return constant('\OCA\SpreedME\Config\Config::' . $key);
+	}
+
+	public static function getDatabaseConfigValue($key) {
+		$value = \OC::$server->getConfig()->getAppValue(Settings::APP_ID, $key);
+		if ($value === 'true' || $value === 'false') {
+			// TODO(leon): How can we improve this?
+			$value = ($value === 'true');
+		}
+		return $value;
+	}
+
+	public static function getDatabaseConfigValueOrDefault($key) {
+		$defaultConfig = array(
+			'SPREED_WEBRTC_ORIGIN' => '',
+			'SPREED_WEBRTC_BASEPATH' => '/webrtc/',
+			'SPREED_WEBRTC_IS_SHARED_INSTANCE' => false,
+			'OWNCLOUD_TEMPORARY_PASSWORD_LOGIN_ENABLED' => false,
+		);
+		if (self::getDatabaseConfigValue('is_set_up') === true) {
+			return self::getDatabaseConfigValue($key);
+		}
+		if (isset($defaultConfig[$key])) {
+			return $defaultConfig[$key];
+		}
+		return '';
+	}
+
+	public static function getConfigValue($key) {
+		if (self::doesPhpConfigExist()) {
+			return self::getFileConfigValue($key);
+		}
+		return self::getDatabaseConfigValue($key);
+	}
+
+	private static function setDatabaseConfigValue($key, $value) {
+		\OC::$server->getConfig()->setAppValue(Settings::APP_ID, $key, $value);
+	}
+
+	public static function setDatabaseConfigValueIfEnabled($key, $value) {
+		if (self::doesPhpConfigExist()) {
+			throw new \Exception('config/config.php exists. Can\'t modify DB config values', ErrorCodes::DB_CONFIG_ERROR_CONFIG_PHP_EXISTS);
+		}
+		self::setDatabaseConfigValue($key, $value);
+	}
+
 	public static function getOwnAppVersion() {
 		return \OCP\App::getAppVersion(Settings::APP_ID);
 	}
@@ -58,13 +105,21 @@ class Helper {
 	}
 
 	public static function notifyIfAppNotSetUp() {
-		if (!class_exists('\OCA\SpreedME\Config\Config', true) || !is_file(self::getOwnAppPath() . 'extra/static/config/OwnCloudConfig.js')) {
-			die('You didn\'t set up this Nextcloud app. Please follow the instructions in the README.md file in the apps/' . Settings::APP_ID . ' folder.');
+		if (!self::doesPhpConfigExist() && self::getDatabaseConfigValue('is_set_up') !== true) {
+			die('You didn\'t set up this Nextcloud app. Please open the Nextcloud admin settings page and configure this app.');
 		}
 	}
 
+	public static function doesPhpConfigExist() {
+		return class_exists('\OCA\SpreedME\Config\Config', true);
+	}
+
+	public static function doesJsConfigExist() {
+		return is_file(self::getOwnAppPath() . 'extra/static/config/OwnCloudConfig.js');
+	}
+
 	public static function getSpreedWebRtcOrigin() {
-		$origin = Config::SPREED_WEBRTC_ORIGIN;
+		$origin = self::getConfigValue('SPREED_WEBRTC_ORIGIN');
 		$is_port = !empty($origin) && $origin[0] === ':';
 		$port = null;
 		if ($is_port) {
@@ -76,14 +131,24 @@ class Helper {
 		return $origin;
 	}
 
-	public static function getSpreedWebRtcUrl($debug = null) {
+	public static function getSpreedWebRtcUrl($debug = null, $includeQueryParams = true) {
 		$origin = self::getSpreedWebRtcOrigin();
-		$basepath = Config::SPREED_WEBRTC_BASEPATH;
-
+		$basepath = self::getConfigValue('SPREED_WEBRTC_BASEPATH');
 		$url = $origin . $basepath;
+		$params = array();
+
 		if ($debug !== false) {
 			if ($debug === true || isset($_GET['debug'])) {
-				$url .= '?debug';
+				$params['debug'] = true;
+			}
+		}
+		if ($includeQueryParams) {
+			if (self::doesJsConfigExist()) {
+				$params['load_config_js'] = true;
+			}
+			$query = http_build_query($params);
+			if (!empty($query)) {
+				$url .= '?' . $query;
 			}
 		}